requirements/base.txt
django==5.1.4
psycopg2-binary==2.9.10
python-environ==0.4.54
django-cors-headers==4.6.0
django-jazzmin==3.0.1
drf-yasg==1.21.8
djangorestframework==3.15.2
django-filter==24.3
django-rosetta==0.10.1
python-telegram-bot==21.9
django-celery-beat==2.7.0
django-debug-toolbar==4.4.6
setuptools
redis==5.2.1
django-celery-results==2.5.1
whitenoise
elasticsearch==8.9.0
elasticsearch-dsl==8.9.0
django-elasticsearch-dsl==8.1
tika==2.6.0
python-dotenv==1.0.1
python-magic==0.4.27